rendered select tag to DOM with data populated
