+++ scripts/convictions/Conviction.js
@@ -1,0 +1,20 @@
+
+
+
+
+
+
+
+
+
+// Reference:
+// export const Conviction = (conviction) => {
+//     return `<section class="conviction__card" id="${conviction.id}"
+//                 <div class="conviction__name">${conviction.name}<div>
+//                 <div class="conviction__age">${conviction.age}<div>
+//                 <div class="conviction__">${conviction.conviction}<div>
+//                 <div class="conviction__termStart">${new Date(conviction.incarceration.start).toLocaleDateString('en-US')}
+//                 <div class="conviction__termEnd">${new Date(conviction.incarceration.end).toLocaleDateString('en-US')}
+//             </section>`
+// }
+
