+++ scripts/convictions/ConvictionProvider.js
@@ -1,0 +1,14 @@
+let convictions = []
+
+export const useConvictions = () => convictions.slice()
+
+export const getConvictions = () => {
+    return fetch("https://criminals.glassdale.us/crimes")
+        .then(response => response.json())
+        .then(
+            parsedConvictions => {
+                console.table(parsedConvictions)
+                convictions = parsedConvictions
+            }
+        )
+}
