+++ scripts/main.js
@@ -1,3 +1,7 @@
+import { ConvictionSelect } from "./convictions/ConvictionSelect.js";
+import { getConvictions } from "./convictions/ConvictionProvider.js";
 import { CriminalList } from "./criminals/CriminalList.js";
 
 CriminalList()
+getConvictions()
+ConvictionSelect()
